use MapViewer class in place viewer
